refactor(edge/benchmark): split benchmark module to connection lookup, route lookup and helper modules
